Clean up & use MDAnalysis as a fallback reader.
Cleaned up some deprecated functions in __init__.py.
Changed internal mechanism of opening reader and topology:
The reader module hanldes both files now, with one functions
that returns Atom and Reader object. When pygmx is not available,
or filetypes can not handled by it, the mdanalysis package can
be used as a fallback now.
The checksum of a reader does not change when mdanalysis is used.